refactor: move files into different folders and update paths
