+++ .github/workflows/build-status-check.yml
@@ -28,4 +28,4 @@ jobs:
         deno-version: ${{ vars.DENO_VERSION }}
 
     - name: Run Build
-      run: deno run --allow-read --allow-run ./.github/cicd/deno-check.ts
+      run: deno run --allow-read --allow-run ./.github/cicd/scripts/deno-check.ts
