+++ .github/workflows/release.yml
@@ -63,7 +63,7 @@ jobs:
           deno-version: ${{ vars.DENO_VERSION }}
 
       - name: Run Build
-        run: deno run --allow-read --allow-run ./.github/cicd/deno-check.ts
+        run: deno run --allow-read --allow-run ./.github/cicd/scripts/deno-check.ts
 
       - name: Run Lint
         run: deno lint
